Refactor test support to accommodate messages controller tests
