+++ app/tests/support/mocks/clients.py
@@ -51,13 +51,16 @@ class AppClientSpec(ClientMockSpec):
     async def edit_username(self, **options):
         await self.send({**RequestSpecs.edit_username(), **options})
         return await self.receive()
-    
+
     async def change_user_avatar(self, **options):
         content: bytes = options.pop('content')
         await self.send({**RequestSpecs.change_user_avatar(), **options})
         self.writer.write(content)
         await self.writer.drain()
         return await self.receive()
+
+    async def new_message(self, **kwargs):
+        await self.send(RequestSpecs.new_message(**kwargs))
 
 class ConnectivityClientSpec(ClientMockSpec):
     async def connect(self, host='localhost', port=8080):
